Some refactoring

+++ platform/x86_64/processor.c
@@ -1,7 +1,7 @@
 #include "processor.h"
 
 extern inline uint64_t x86_64_PageEntry_GetAddr(x86_64_PageEntry entry);
-extern inline int x86_64_PageEntry_GetFlags(x86_64_PageEntry entry);
+extern inline uint64_t x86_64_PageEntry_GetFlags(x86_64_PageEntry entry);
 extern inline int x86_64_PageEntry_GetProtectionKey(x86_64_PageEntry entry);
 extern inline void x86_64_FlushPageTLB(volatile void* addr);
 extern inline void* x86_64_FlushPageTLB2(uintptr_t addr);
